requirements.txt
flask
Flask
flask-cors
flask-login
flask-migrate
flask_seeder
SQLAlchemy-serializer
gunicorn
requests
beautifulsoup4
psycopg2
psycopg2-binary
bcrypt
python-dotenv
pytest
